[UXE-5949] chore: update Azion CLI version and adjust deployment commands for workflows (#2173)
* chore: update Azion CLI version and adjust deployment commands for production and staging workflows

* chore: update Azion configuration by adding new dependency and removing obsolete config file

* chore: update Azion CLI version to 2.5.3 in production deployment workflow
