+++ package.json
@@ -40,6 +40,7 @@
     "@vueuse/core": "^10.7.0",
     "analytics": "^0.8.11",
     "axios": "^1.7.4",
+    "azion": "^1.15.0",
     "azion-theme": "^1.8.0",
     "c3": "^0.7.20",
     "deep-chat": "^1.4.11",
